geca_calendar/requirements.txt
arrow==1.3.0
attrs==24.1.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
gunicorn==22.0.0
ics==0.7.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
TatSu==5.8.3
types-python-dateutil==2.9.0.20240316
urllib3==2.2.2
Werkzeug==3.0.3
